indicate event for a dead gene
