+++ client/src/containers/Gene/GeneActivitiesTable.js
@@ -14,7 +14,7 @@ import {
   Timestamp,
   Typography,
 } from '../../components/elements';
-import { pastTense } from '../../utils/events';
+import { pastTense, getActivityDescriptor } from '../../utils/events';
 import ResurrectGeneDialog from './ResurrectGeneDialog';
 import UndoMergeGeneDialog from './UndoMergeGeneDialog';
 import UndoSplitGeneDialog from './UndoSplitGeneDialog';
@@ -77,61 +77,6 @@ class GeneActivitiesTable extends Component {
         ) : null}
       </div>
     );
-  };
-
-  getActivityDescriptor = (activityItem = {}, selfGeneId) => {
-    const what = activityItem['provenance/what'];
-    const { statusChange, relatedGeneId } = (activityItem.changes || []).reduce(
-      (result, change) => {
-        const { attr, value, added } = change || {};
-        if (added)
-          if (attr === 'gene/splits' || attr === 'gene/merges') {
-            return {
-              ...result,
-              relatedGeneId: value,
-            };
-          } else if (attr === 'gene/status') {
-            return {
-              ...result,
-              statusChange: value,
-            };
-          }
-        return result;
-      },
-      {}
-    );
-
-    let eventType;
-    if (what === 'event/merge-genes' && !statusChange) {
-      eventType = 'merge_from';
-    } else if (
-      what === 'event/merge-genes' &&
-      statusChange === 'gene.status/dead'
-    ) {
-      eventType = 'merge_into';
-    } else if (what === 'event/split-gene' && !statusChange) {
-      eventType = 'split_into';
-    } else if (
-      what === 'event/split-gene' &&
-      statusChange === 'gene.status/live'
-    ) {
-      eventType = 'split_from';
-    } else {
-      eventType = what;
-    }
-
-    const descriptor = {
-      eventLabel: eventType || activityItem['provenance/what'],
-      entity: {
-        'gene/id': selfGeneId,
-      },
-      relatedEntity: relatedGeneId
-        ? {
-            'gene/id': relatedGeneId,
-          }
-        : null,
-    };
-    return descriptor;
   };
 
   getGeneIdForEvent = (selectedActivity, eventType) => {
@@ -205,10 +150,7 @@ class GeneActivitiesTable extends Component {
                 eventLabel,
                 entity,
                 relatedEntity,
-              } = this.getActivityDescriptor(
-                activityItem,
-                this.props.selfGeneId
-              );
+              } = getActivityDescriptor(activityItem, this.props.selfGeneId);
               return (
                 <TableRow key={activityIndex}>
                   <TableCell className={classes.time}>
